+++ NEWS.md
@@ -30,6 +30,10 @@ Newest versions at top.
  * Handle multiple phenotypes using in the PLINK way, using `--all-pheno`
  * Add `save_qassoc_table_to_file` to save the results of a quantitive
    association to file in the same way PLINK creates a `.qassoc` file
+ * Renamed column names `family_id` to `FID`, as this is the abbreviation
+   PLINK uses, see https://www.cog-genomics.org/plink/1.9/input#pheno
+ * Renamed column names `within_family_id` to `IID`, as this is the abbreviation
+   PLINK uses, see https://www.cog-genomics.org/plink/1.9/input#pheno
 
 ### MINOR IMPROVEMENTS